feature: accommodation 예약쿠폰적용 및 카드혜택적용

+++ src/main/java/org/cardGGaduekMainService/coupon/memberCoupon/domain/MemberCouponVO.java
@@ -21,7 +21,7 @@ public class MemberCouponVO {
     private Long couponProductId;
     private String couponName;
     private String couponCategory;
-    private Long discountValue;
+    private BigDecimal discountValue;
 
     private String couponStatus;     // 사용 가능, 사용 완료
     private LocalDateTime expiredAt;
